requirements.txt
flask>=3.0.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
youtube-transcript-api==0.6.2
openai>=1.30.0
anthropic>=0.34.0
python-dotenv==1.0.1
requests==2.32.3
